Regular push

+++ app/src/main/java/com/appdev/split/Model/Data/PaymentDistribute.kt
@@ -3,5 +3,5 @@ package com.appdev.split.Model.Data
 data class PaymentDistribute(
     val id: String,
     val name: String,
-    var amount: Float = 0f
+    var amount: Double = 0.0
 )
