+++ app/src/main/java/com/appdev/split/Model/Data/Percentage.kt
@@ -3,6 +3,6 @@ package com.appdev.split.Model.Data
 data class Percentage(
     val id: String,
     val name: String,
-    var percentage: Float = 0f,
-    var amount: Float = 0f
+    var percentage: Double = 0.0,
+    var amount: Double = 0.0
 )
